Move impl Contains for NetAddr into netaddr mod
Tested-by: Kristofer Rye <kristofer.rye@gmail.com>
